+++ wagtail/core/models/__init__.py
@@ -85,16 +85,6 @@ from .audit_log import (  # noqa
     LogEntryQuerySet,
     ModelLogEntry,
 )
-from .collections import (  # noqa
-    BaseCollectionManager,
-    Collection,
-    CollectionManager,
-    CollectionMember,
-    CollectionViewRestriction,
-    GroupCollectionPermission,
-    GroupCollectionPermissionManager,
-    get_root_collection_id,
-)
 from .copying import _copy, _copy_m2m_relations, _extract_field_data  # noqa
 from .i18n import (  # noqa
     BootstrapTranslatableMixin,
@@ -104,6 +94,16 @@ from .i18n import (  # noqa
     TranslatableMixin,
     bootstrap_translatable_model,
     get_translatable_models,
+)
+from .media import (  # noqa
+    BaseCollectionManager,
+    Collection,
+    CollectionManager,
+    CollectionMember,
+    CollectionViewRestriction,
+    GroupCollectionPermission,
+    GroupCollectionPermissionManager,
+    get_root_collection_id,
 )
 from .sites import Site, SiteManager, SiteRootPath  # noqa
 from .view_restrictions import BaseViewRestriction
