+++ wagtail/core/models/media.py
